BE/requirements.txt
tensorflow==2.15.0
opencv-python==4.8.0.76
numpy==1.26.4
scikit-learn==1.3.2
pandas==2.2.2
psycopg2-binary==2.9.9
pgvector==0.2.5
faiss-cpu==1.8.0
flask==3.0.3
flask-cors
python-dotenv
Pillow>=9.0.0
